requirements.txt
### requirements.txt
openai==0.28
pillow
tkinter
requests